backend/requirements.txt
Django
gunicorn
psycopg2
django-environ
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-redis
